fix: Hide checkbox for text messages
DMS does not want to deal with the effort needed for sending SMS at this
time, and this checkbox confuses people/gives them a false sense.

It is being hidden, because we might want it back at some point, and I
didn't want to make a DB change to set this field non required.